requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
chex==0.1.90
contourpy==1.3.2
cycler==0.12.1
dm-haiku==0.0.14
etils==1.13.0
filelock==3.19.1
flax==0.10.7
fonttools==4.61.0
fsspec==2025.9.0
humanize==4.14.0
importlib_resources==6.5.2
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
Jinja2==3.1.3
jmp==0.0.4
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.30
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.2
psutil==7.1.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
scikit-learn==1.5.0
scipy==1.15.3
seaborn==0.13.2
simplejson==3.20.2
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tensorstore==0.1.78
threadpoolctl==3.6.0
toolz==1.1.0
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchvision==0.24.1+cu128
tqdm==4.66.5
treescope==0.1.10
triton==3.5.1
typing_extensions==4.12.2
tzdata==2025.2
zipp==3.23.0
